fix: race conditions in stream creation
`do_send_headers/3` is now "serialized" by writing all data to the
socket before sending the next set of headers. This is slower, but fixes
all of the race conditions that were cropping up (stream ID's out of
order is a `PROTOCOL_ERROR`). Perhaps theres a way to make this faster?