implement a configurable superivision tree
Currently Kadabra starts an application which is problematic
in some failure scenarios that can result in the entire BEAM node
being restarted.
Starting an application in a library doesn't put the library user in control.
This fixes it in a backwards compatible way by allowing the library user
to decide exactly how their application should behave when Kadabra
crashes.